requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.4
dj-database-url==1.2.0
Django==4.1.4
django-bootstrap-v5==1.0.11
django-fontawesome-6==1.0.0.0
django-heroku==0.3.1
django-htmx==1.13.0
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
lxml==4.9.2
multidict==6.0.4
ntlm-auth==1.5.0
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
python-decouple==3.6
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.10.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
yarl==1.8.2
zeep==4.2.1
